:recycle: Moved build run from infra to container-registry
