+++ apps/container-registry/Taskfile.yml
@@ -38,6 +38,7 @@ tasks:
           --struct github.com/kloudlite/api/apps/container-registry/internal/domain/entities.GithubListRepository
           --struct github.com/kloudlite/api/apps/container-registry/internal/domain/entities.GithubSearchRepository
           --struct github.com/kloudlite/api/apps/container-registry/internal/domain/entities.GitBranch
+          --struct github.com/kloudlite/api/apps/container-registry/internal/domain/entities.BuildRun          
           --struct github.com/kloudlite/api/pkg/types.Pagination
           --struct github.com/kloudlite/api/apps/container-registry/internal/domain/entities.Credential
           --struct github.com/kloudlite/api/apps/container-registry/internal/domain/entities.Repository
@@ -49,7 +50,7 @@ tasks:
         > ./internal/app/_struct-to-graphql/main.go
       - |+
         pushd ./internal/app/_struct-to-graphql
-        go run main.go --dev --out-dir ../graph/struct-to-graphql --with-pagination Credential,Repository,Digest,Build,BuildCacheKey
+        go run main.go --dev --out-dir ../graph/struct-to-graphql --with-pagination Credential,BuildRun,Repository,Digest,Build,BuildCacheKey
         popd
       - rm -rf ./internal/app/_struct-to-graphql
