+++ apps/container-registry/internal/app/graph/struct-to-graphql/buildrun.graphqls
