+++ apps/container-registry/internal/domain/api.go
@@ -77,5 +77,10 @@ type Domain interface {
 	DeleteBuildCache(ctx RegistryContext, id repos.ID) error
 	ListBuildCaches(ctx RegistryContext, search map[string]repos.MatchFilter, pagination repos.CursorPagination) (*repos.PaginatedRecord[*entities.BuildCacheKey], error)
 
+	ListBuildRuns(ctx RegistryContext, repoName string, search map[string]repos.MatchFilter, pagination repos.CursorPagination) (*repos.PaginatedRecord[*entities.BuildRun], error)
+	GetBuildRun(ctx RegistryContext, repoName string, runName string) (*entities.BuildRun, error)
+	OnBuildRunUpdateMessage(ctx RegistryContext, clusterName string, buildRun entities.BuildRun) error
+	OnBuildRunDeleteMessage(ctx RegistryContext, clusterName string, buildRun entities.BuildRun) error
+
 	ListBuildsByCache(ctx RegistryContext, cacheId repos.ID, pagination repos.CursorPagination) (*repos.PaginatedRecord[*entities.Build], error)
 }
