+++ apps/infra/Taskfile.yml
@@ -40,14 +40,13 @@ tasks:
           --struct github.com/kloudlite/api/apps/infra/internal/entities.DomainEntry
           --struct github.com/kloudlite/api/apps/infra/internal/entities.VPNDevice
           --struct github.com/kloudlite/api/apps/infra/internal/entities.PersistentVolumeClaim
-          --struct github.com/kloudlite/api/apps/infra/internal/entities.BuildRun
           --struct github.com/kloudlite/api/apps/infra/internal/entities.ClusterManagedService
           --struct github.com/kloudlite/api/pkg/repos.MatchFilter
           --struct github.com/kloudlite/api/pkg/repos.CursorPagination
         > ./internal/app/_struct-to-graphql/main.go
       - |+
         pushd ./internal/app/_struct-to-graphql
-        go run main.go --dev --out-dir ../graph/struct-to-graphql --with-pagination Cluster,Node,NodePool,CloudProviderSecret,DomainEntry,VPNDevice,BuildRun,PersistentVolumeClaim,ClusterManagedService
+        go run main.go --dev --out-dir ../graph/struct-to-graphql --with-pagination Cluster,Node,NodePool,CloudProviderSecret,DomainEntry,VPNDevice,PersistentVolumeClaim,ClusterManagedService
         popd
       - rm -rf ./internal/app/_struct-to-graphql
