+++ apps/infra/internal/domain/api.go
@@ -80,11 +80,6 @@ type Domain interface {
 	OnPVCUpdateMessage(ctx InfraContext, clusterName string, pvc entities.PersistentVolumeClaim) error
 	OnPVCDeleteMessage(ctx InfraContext, clusterName string, pvc entities.PersistentVolumeClaim) error
 
-	ListBuildRuns(ctx InfraContext, repoName string, search map[string]repos.MatchFilter, pagination repos.CursorPagination) (*repos.PaginatedRecord[*entities.BuildRun], error)
-	GetBuildRun(ctx InfraContext, repoName string, runName string) (*entities.BuildRun, error)
-	OnBuildRunUpdateMessage(ctx InfraContext, clusterName string, buildRun entities.BuildRun) error
-	OnBuildRunDeleteMessage(ctx InfraContext, clusterName string, buildRun entities.BuildRun) error
-
 	ListClusterManagedServices(ctx InfraContext, clusterName string, search map[string]repos.MatchFilter, pagination repos.CursorPagination) (*repos.PaginatedRecord[*entities.ClusterManagedService], error)
 	GetClusterManagedService(ctx InfraContext, clusterName string, serviceName string) (*entities.ClusterManagedService, error)
 	CreateClusterManagedService(ctx InfraContext, clusterName string, service entities.ClusterManagedService) (*entities.ClusterManagedService, error)
